requirements.txt
sqlalchemy
psycopg2-binary
selenium
python-dotenv
